Create new `sys_ops` crate
This moves sys op definitions from `sys_types` into a new crate to avoid cyclic dependencies.
