+++ baml_language/Cargo.toml
@@ -45,6 +45,7 @@ baml_compiler_syntax = { path = "crates/baml_compiler_syntax" }
 bex_project = { path = "crates/bex_project" }
 bex_engine = { path = "crates/bex_engine" }
 sys_llm = { path = "crates/sys_llm" }
+sys_ops = { path = "crates/sys_ops" }
 sys_types = { path = "crates/sys_types" }
 sys_native = { path = "crates/sys_native" }
 baml_db = { path = "crates/baml_db" }
