+++ baml_language/crates/baml_lsp_server/Cargo.toml
@@ -17,6 +17,7 @@ bex_heap = { workspace = true }
 bex_project = { workspace = true }
 bridge_ctypes = { workspace = true }
 sys_native = { workspace = true, features = [ "bundle-http" ] }
+sys_ops = { workspace = true }
 sys_types = { workspace = true }
 anyhow = { workspace = true }
 axum = { workspace = true }
