+++ baml_language/crates/baml_tests/Cargo.toml
@@ -26,7 +26,8 @@ baml_workspace = { workspace = true }
 bex_engine = { workspace = true }
 bex_vm = { workspace = true }
 bex_vm_types = { workspace = true }
-sys_native = { workspace = true, features = [ "bundle-http" ] }
+sys_native = { workspace = true, features = ["bundle-http"] }
+sys_ops = { workspace = true }
 sys_types = { workspace = true }
 anyhow = { workspace = true }
 indexmap = { workspace = true }
@@ -47,7 +48,7 @@ baml_compiler2_emit = { workspace = true }
 bex_external_types = { workspace = true }
 divan = { workspace = true }
 tempfile = { workspace = true }
-tokio = { workspace = true, features = [ "rt-multi-thread", "macros", "net" ] }
+tokio = { workspace = true, features = ["rt-multi-thread", "macros", "net"] }
 wiremock = { workspace = true }
 
 [[bench]]
@@ -55,8 +56,8 @@ name = "compiler_benchmark"
 harness = false
 
 [features]
-default = [  ]
-heap_debug = [ "bex_vm/heap_debug", "bex_vm_types/heap_debug" ]
+default = []
+heap_debug = ["bex_vm/heap_debug", "bex_vm_types/heap_debug"]
 
 [package.metadata.ci]
 wasm_support = false
