+++ baml_language/crates/baml_tests/src/engine.rs
@@ -211,7 +211,7 @@ pub async fn run_test(
     let positional_args = resolve_args(&program, entry, args);
 
     // Create engine and execute.
-    let engine = BexEngine::new(program, Arc::new(sys_types::SysOps::native()), None)
+    let engine = BexEngine::new(program, Arc::new(sys_ops::SysOps::native()), None)
         .expect("Failed to create BexEngine");
     let engine = Arc::new(engine);
