+++ baml_language/crates/bex_engine/tests/build_request.rs
@@ -32,7 +32,7 @@ async fn run_baml_with_args(
 ) -> BexExternalValue {
     let snapshot = common::compile_for_engine(source);
     let engine = Arc::new(
-        BexEngine::new(snapshot, sys_types::SysOps::native().into(), None)
+        BexEngine::new(snapshot, sys_native::SysOps::native().into(), None)
             .expect("Failed to create engine"),
     );
