+++ baml_language/crates/bex_engine/tests/common/mod.rs
@@ -73,7 +73,7 @@ pub(crate) async fn assert_engine_executes(input: EngineProgram) -> anyhow::Resu
 
     let snapshot = compile_for_engine(&source);
     let engine = Arc::new(
-        BexEngine::new(snapshot, Arc::new(sys_types::SysOps::native()), None)
+        BexEngine::new(snapshot, Arc::new(sys_native::SysOps::native()), None)
             .expect("Failed to create engine"),
     );
 
@@ -201,7 +201,7 @@ function get_prompt() -> baml.llm.PromptAst {{
     for i in 0..3 {
         let snapshot = compile_for_engine(&source);
         let engine = Arc::new(
-            BexEngine::new(snapshot, Arc::new(sys_types::SysOps::native()), None)
+            BexEngine::new(snapshot, Arc::new(sys_native::SysOps::native()), None)
                 .expect("Failed to create engine"),
         );
