+++ baml_language/crates/bex_engine/tests/llm_render.rs
@@ -243,7 +243,7 @@ function test_render() -> int {
 
     let snapshot = common::compile_for_engine(source);
     let engine = std::sync::Arc::new(
-        BexEngine::new(snapshot, sys_types::SysOps::native().into(), None)
+        BexEngine::new(snapshot, sys_native::SysOps::native().into(), None)
             .expect("Failed to create engine"),
     );
 
@@ -299,7 +299,7 @@ function get_prompt() -> baml.llm.PromptAst {
 
     let snapshot = common::compile_for_engine(source);
     let engine = std::sync::Arc::new(
-        BexEngine::new(snapshot, sys_types::SysOps::native().into(), None)
+        BexEngine::new(snapshot, sys_native::SysOps::native().into(), None)
             .expect("Failed to create engine"),
     );
 
@@ -376,7 +376,7 @@ function test_build_request() -> int {
 
     let snapshot = common::compile_for_engine(source);
     let engine = std::sync::Arc::new(
-        BexEngine::new(snapshot, sys_types::SysOps::native().into(), None)
+        BexEngine::new(snapshot, sys_native::SysOps::native().into(), None)
             .expect("Failed to create engine"),
     );
 
@@ -418,7 +418,7 @@ function test_call_llm() -> unknown {
 
     let snapshot = common::compile_for_engine(source);
     let engine = std::sync::Arc::new(
-        BexEngine::new(snapshot, sys_types::SysOps::native().into(), None)
+        BexEngine::new(snapshot, sys_native::SysOps::native().into(), None)
             .expect("Failed to create engine"),
     );
 
@@ -466,7 +466,7 @@ function test_call_llm() -> string {
 
     let snapshot = common::compile_for_engine(source);
     let engine = std::sync::Arc::new(
-        BexEngine::new(snapshot, sys_types::SysOps::native().into(), None)
+        BexEngine::new(snapshot, sys_native::SysOps::native().into(), None)
             .expect("Failed to create engine"),
     );
 
@@ -515,7 +515,7 @@ function test_call_llm() -> unknown {
 
     let snapshot = common::compile_for_engine(source);
     let engine = std::sync::Arc::new(
-        BexEngine::new(snapshot, sys_types::SysOps::native().into(), None)
+        BexEngine::new(snapshot, sys_native::SysOps::native().into(), None)
             .expect("Failed to create engine"),
     );
 
@@ -591,7 +591,7 @@ function get_prompt() -> baml.llm.PromptAst {
 
     let snapshot = common::compile_for_engine(source);
     let engine = std::sync::Arc::new(
-        BexEngine::new(snapshot, sys_types::SysOps::native().into(), None)
+        BexEngine::new(snapshot, sys_native::SysOps::native().into(), None)
             .expect("Failed to create engine"),
     );
 
@@ -636,7 +636,7 @@ function get_prompt() -> baml.llm.PromptAst {
 
     let snapshot = common::compile_for_engine(source);
     let engine = std::sync::Arc::new(
-        BexEngine::new(snapshot, sys_types::SysOps::native().into(), None)
+        BexEngine::new(snapshot, sys_native::SysOps::native().into(), None)
             .expect("Failed to create engine"),
     );
 
@@ -687,7 +687,7 @@ function get_prompt() -> baml.llm.PromptAst {
 
     let snapshot = common::compile_for_engine(source);
     let engine = std::sync::Arc::new(
-        BexEngine::new(snapshot, sys_types::SysOps::native().into(), None)
+        BexEngine::new(snapshot, sys_native::SysOps::native().into(), None)
             .expect("Failed to create engine"),
     );
 
@@ -732,7 +732,7 @@ function get_prompt() -> baml.llm.PromptAst {
 
     let snapshot = common::compile_for_engine(source);
     let engine = std::sync::Arc::new(
-        BexEngine::new(snapshot, sys_types::SysOps::native().into(), None)
+        BexEngine::new(snapshot, sys_native::SysOps::native().into(), None)
             .expect("Failed to create engine"),
     );
