+++ baml_language/crates/bex_project/Cargo.toml
@@ -33,6 +33,7 @@ bex_events = { workspace = true }
 bex_external_types = { workspace = true }
 bex_heap = { workspace = true }
 bex_vm_types = { workspace = true }
+sys_ops = { workspace = true }
 sys_types = { workspace = true }
 async-trait = { workspace = true }
 crossbeam = { workspace = true }
