+++ baml_language/crates/bex_project/src/project.rs
@@ -1,5 +1,5 @@
 use bex_engine::BexEngine;
-use sys_types::SysOps;
+use sys_ops::SysOps;
 
 use crate::RuntimeError;
