+++ baml_language/crates/bridge_wasm/Cargo.toml
@@ -15,6 +15,7 @@ baml_type = { workspace = true }
 bex_heap = { workspace = true }
 bex_project = { workspace = true }
 bridge_ctypes = { workspace = true }
+sys_ops = { workspace = true }
 sys_types = { workspace = true }
 async-trait = { workspace = true }
 console_error_panic_hook = { workspace = true, optional = true }
