+++ baml_language/crates/bridge_wasm/src/wasm_env.rs
@@ -7,7 +7,8 @@
 use std::sync::Arc;
 
 use js_sys::{Function, Promise};
-use sys_types::{BexHeap, CallId, OpErrorKind, SysOpContext, SysOpOutput, io::IoNamespaceEnv};
+use sys_ops::io::IoNamespaceEnv;
+use sys_types::{BexHeap, CallId, OpErrorKind, SysOpContext, SysOpOutput};
 use wasm_bindgen::JsCast;
 use wasm_bindgen_futures::JsFuture;
