+++ baml_language/crates/bridge_wasm/src/wasm_sys.rs
@@ -1,6 +1,7 @@
 use std::sync::Arc;
 
-use sys_types::{BexHeap, CallId, OpErrorKind, SysOpContext, SysOpOutput, io::IoNamespaceSys};
+use sys_ops::io::IoNamespaceSys;
+use sys_types::{BexHeap, CallId, OpErrorKind, SysOpContext, SysOpOutput};
 use wasm_bindgen::prelude::*;
 
 use crate::send_wrapper::SendFuture;
